fix: カードのアニメーションの記載場所を移動しました
